WIP
So the code get executed when you click on the ugly hacky button I added (import geojson)

It calls the function "import_csv_fun" in sidebar.js which make a GET request to /import-csv url,
That return the csv in a jsonified form with values are in a single array. Because I know the column order,
and the length N of one column (i.e number of rows), I can loop through 1 --> N and rebuild each feature by
assigning the values to the correct property. First column property will get value at index 0, then index 0 + N,
the index 0 + 2*N etc... The jth column will have value at index j + i*N, where i is the row number (between 0 and N-1) (a table is a
matrix with rows and columns and it can be mapped to an array of N * M values). Hope it make some sense

It get quick because we now make only one for loop with N steps and not a nested loop of N*M steps